moved RangeSlider class to new file for clarity and modularity
